Update set-up-ipsec-vpn.md
Updated per conversation with Arun.  

We are adding language to guide a customer to another product selection if their use case is not specifically backend management.  We also want to discourage any attempts to work with other cloud providers through this service as it is not defined to do so.  Customers must use other products for this type of VPN.